backend/requirements.txt
anyio==3.7.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
httptools==0.5.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
python-multipart
pydantic[email]
